Move models from mining api entry to models dedicated package
